- Adjusted some algorithms so that they can handle changes in the input image size - Various other minor changes - Refactored a fuction from getCorners() to getFeatures() - Added and tweaked scale space feature detectors

+++ ip/src/gecv/core/image/border/FactoryImageBorder.java
@@ -135,7 +135,7 @@ public class FactoryImageBorder {
 		return ret;
 	}
 
-	public static ImageBorder_F64 value( ImageFloat64 image , float value ) {
+	public static ImageBorder_F64 value( ImageFloat64 image , double value ) {
 		return ImageBorderValue.wrap(image,value);
 	}
